+++ lib/odd_job/pool/supervisor.ex
@@ -4,15 +4,14 @@ defmodule OddJob.Pool.Supervisor do
 
   use Supervisor
 
-  def start_link(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: id(opts[:name]))
+  def start_link([name, _opts] = args) do
+    Supervisor.start_link(__MODULE__, args, name: id(name))
   end
 
-  def init(opts) do
+  def init([name, opts]) do
     config = Application.get_all_env(:odd_job)
     default_pool_size = Keyword.get(config, :pool_size, 5)
     {pool_size, opts} = Keyword.pop(opts, :pool_size, default_pool_size)
-    {name, opts} = Keyword.pop!(opts, :name)
 
     default_opts = [
       strategy: :one_for_one,
@@ -30,9 +29,9 @@ defmodule OddJob.Pool.Supervisor do
 
   defp workers(name, pool_size) do
     for num <- 1..pool_size do
-      id = "#{name}_worker_#{num}"
+      id = :"#{name}_worker_#{num}"
       pool_id = OddJob.pool_id(name)
-      {OddJob.Pool.Worker, %{id: id, pool: name, pool_id: pool_id}}
+      {OddJob.Pool.Worker, id: id, pool: name, pool_id: pool_id}
     end
   end
 end