+++ lib/odd_job/supervisor.ex
@@ -9,59 +9,47 @@ defmodule OddJob.Supervisor do
   @moduledoc since: "0.1.0"
   use Supervisor
 
-  @type start_arg :: atom | [start_option]
+  @type start_arg :: atom | [{:name, atom} | start_option]
+  @type child_spec :: Supervisor.child_spec()
   @type start_option ::
-          {:name, atom}
-          | {:pool_size, non_neg_integer}
+          {:pool_size, non_neg_integer}
           | {:max_restarts, non_neg_integer}
           | {:max_seconds, non_neg_integer}
-  @type child_spec :: %{
-          id: atom,
-          start: {OddJob.Supervisor, :start_link, [start_option]},
-          type: :supervisor
-        }
 
   @doc false
-  @spec start_link(start_arg) :: Supervisor.on_start()
-  def start_link(start_arg)
-
-  def start_link(name) when is_atom(name) do
-    start_link(name: name)
-  end
-
-  def start_link(opts) when is_list(opts) do
-    Supervisor.start_link(__MODULE__, opts, name: id(opts[:name]))
+  @spec start_link(atom, [start_option]) :: Supervisor.on_start()
+  def start_link(name, opts \\ []) when is_atom(name) and is_list(opts) do
+    opts = Keyword.delete(opts, :name)
+    Supervisor.start_link(__MODULE__, [name, opts], name: id(name))
   end
 
   @impl true
-  def init(opts) do
-    name = Keyword.get(opts, :name)
-    pool_opts = %{id: pool_id(name), pool: name}
+  def init([name, _opts] = args) do
+    pool_opts = [id: pool_id(name), pool: name]
 
     children = [
       {DynamicSupervisor, strategy: :one_for_one, name: pool_id(name) |> proxy_sup_name()},
       {DynamicSupervisor, strategy: :one_for_one, name: scheduler_sup_name(name)},
       {OddJob.Pool, pool_opts},
-      {OddJob.Pool.Supervisor, opts}
+      {OddJob.Pool.Supervisor, args}
     ]
 
     Supervisor.init(children, strategy: :one_for_one)
   end
 
   @doc false
-  @spec child_spec(start_arg | {atom, [start_option]}) :: child_spec
+  @spec child_spec(start_arg) :: child_spec
   def child_spec(name) when is_atom(name), do: child_spec(name: name)
 
-  def child_spec({name, opts}) when is_atom(name) and is_list(opts) do
-    opts
-    |> Keyword.put(:name, name)
-    |> child_spec()
-  end
-
   def child_spec(opts) when is_list(opts) do
+    {name, opts} = Keyword.pop!(opts, :name)
+
     opts
     |> super()
-    |> Supervisor.child_spec(id: id(opts[:name]))
+    |> Supervisor.child_spec(
+      id: id(name),
+      start: {__MODULE__, :start_link, [name, opts]}
+    )
   end
 
   @doc false